Services catalog with templates #89
First version :
When creating/updating a SC, it is possible to define a template
parameter. When SC is a template, a SC is created in DB for every sub
entity. All sub SCs parameters are copied/updated from master SC.
When Shinken configuration is built a business rule is built for every
sub SC.

Known issue : it is mandatory to save SC twice when setting the new
template parameter ... do not understand why !