Set current user Alignak backend in the session
If token is not in the session, log in to the Alignak backend (with
current user configured username/password or token) and set the backend
token in the PHP session.
If token is in the PHP session, use this token and do not log in the
backend.
Return the token to the function caller.